feat: refatora funcoes
